streamlit_requirements.txt
streamlit==1.32.0
pytube==12.1.3
google-api-python-client==2.107.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Pillow==10.2.0
streamlit-authenticator==0.2.3
pyyaml==6.0.1
watchdog==3.0.0 